10-88
Fixed issue when there were no references to crosshair ui.

Weapon impact effect separated as method.

Weapon impact is now replicated as well.

Still, not a fan of how its handled, if the structure doesnt change it does not get replicated, and as it is fails to propagate direction in the server version.  If Tom does not fix it i will XD, but for now lets follow the course.