Work in progress 13-126
Game mode now has an event, called from health component, that allows to broadcast data whenever a player gets killed. We bind to this event in the gamemode BP so we can increase the score in the gamestate and prnt it.

Currently it asumes that the victim is always a pawn, we need to correct that