14-133
Creating EQS to find cover:

A good cover  is a place as close as possible to the querier, where the cover point and the player cant see (score influence only criteria, because this is a suggestion not an absolute, worse score for cover places that the player can see), and discarding high covers (traces with a height offset, if they cant see each other, discard)

Donut, to check points around querier.
a. Test: Distance to querier, with filter type minimum and score -1. Because we want the positions further away in the donut, the worse score, the closest, better score.
b. Test: Trace, with context all players (from a previous lecture. Because we want to know if from our point is possible to see the player). It will give bad score to EQS places the player can see. We change it to test purpose score only, because we want it to be a sugestion, not an order.
c. Test: To see if its low cover or high cover, another trace test, with a context of players too, with an item height offset of 100 and context height offset of 100, because we want to differentiate low covers form high covers.  In this case, have the bool match turned off, because we want the opossite.

Also, for tweaking, lets set the trace mode to navigation, because we dont want the height to matter, the navigation. Also, lets turn on the check on Use spiral pattern, to randomize a little. Distance to querier, the first test, lets put it in -2 because we want it to weight a lot more.

And lets put a lot of more rings and points, for better testing.