14-137
Creating EQS to find nearest player, without sensing it, because when the enemy sees no player we want it to move near the closest player to make everything engaging. In the future it could be used or a patrol, so lets refer to it as "patrol".

The BT for the case of not looking at a player, instead of just waiting, it will move the character to the nearest player. And by setting the "Verify if there is target actor" to "observer aborts lower priority", it will abort said "patrol" in case it looks at a player.

Temporaly desactivating bot spawning in the gamemode