14-138
Reconfigure a little the tree, for the attack.

Create attack task, (start and end fire are now public and blupritnt calleable).  It will begin attack, wait a little bit and the succeed. Only aborts if the AI is being called in anon coopcharacter.

Using a default task, default focus, to have the ai turn to the player as we need.

Important, the decorator "I have a player to attack " will abort lower priotity, because if we found a player, there is no need to "try to patrol".

Ok, now we have a killer machine that will try to engage the closest player, then move to a cover, and then fire to it. Really accurate !