2-3 to 2-4
Adding a player controller.

Adding basic movement and camera orientation

Check in character movement that "Orient Rotation to movement" is on.
Check in Character that "use controller rotation pitch" and use "controller rotation yaw" are off.
Check that in the spring arm "use pawn control rotation" is on

By doing so we enssure that only the camera moves around when using the mouse and that if you move the camera while moving the character will point towards the movement direction.

We made the rotation rate from 360 to 720, so it catches up faster when moving the camera