2-5
Create animation blueprint.
The animation blueprint updates the velocity of the character, and checks if its in the air.

2-6

Working on anim graph!

Adding state for idle, walk and run.

Creating a blend space to easily manage the transition bettwen idle, walk and run states using the paragorn free animations.

In horizontal and vertical axis set number of grid divisions to 8, and interpolation time to 0.1..put minimum and maximum axis values from -300 to 300

Now we have our paragorn character moving around.