2-9  and 2-10
Polishing the attack animation, pretty much by tweaking when the attack animation is full body and when half body.

Now we have a half body version of the ability, going up from one bone near the hips.

What we do is have a notifiy to play only as full body during the actuall full body attack portion, so it looks better.

We use a bool based blend to decide when to play full body or half body ability. We put the time for the false blend from 0.1 to 0.3, so that it takes a little more and looks better.

Also, if the character is still, we play the full body version.

Looking good.