3-12
Starting using an melee ability.

Now we refactor the attack montage calling it from the melee ability. This ability so far only calls the montage. We have also an gameplay effect, that serves us to have the ability cooldown of 1 second.

We added a gameplay tag to the cooldown melee effect but doesnt seem to do anything so far.

Seems a little messy, but very practical to quickly do abilities once we are used to it.