3-14
By brute force, adding a collider attached to the sword socket, so we can start working ina melee attack ability.

Also adding Jade, a Shinbi with other skin that will serve as enemy. I added myself a ery basic AI for Jade, it patrols and goes to the player if notices it.